pyproject.toml
[project]
name = "smart-shop-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "docling>=2.41.0",
    "fastapi>=0.116.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.7",
    "langchain-groq>=0.3.5",
    "langchain-ollama>=0.3.4",
    "langchain-tavily>=0.2.7",
    "langgraph>=0.5.2",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
